rename settings to defaultSettings
