+++ src/sparnatural/components/builder-section/groupwrapper/LinkWhereBottom.ts
@@ -1,4 +1,4 @@
-import { getSettings } from "../../../../configs/client-configs/settings";
+import { getSettings } from "../../../../configs/client-configs/defaultSettings";
 import HTMLComponent from "../../HtmlComponent";
 import EndClassGroup from "./criteriagroup/startendclassgroup/EndClassGroup";
 import GroupWrapper from "./GroupWrapper";
