+++ src/sparnatural/components/buttons/VariableOptionsSelectBtn.ts
@@ -1,4 +1,4 @@
-import { getSettings } from "../../../configs/client-configs/settings";
+import { getSettings } from "../../../configs/client-configs/defaultSettings";
 import HTMLComponent from "../HtmlComponent";
 
 class VariableOptionsSelectBtn extends HTMLComponent {
