+++ src/sparnatural/components/buttons/actions/actioncomponents/ActionAnd.ts
@@ -1,4 +1,4 @@
-import { getSettings } from "../../../../../configs/client-configs/settings";
+import { getSettings } from "../../../../../configs/client-configs/defaultSettings";
 import HTMLComponent from "../../../HtmlComponent";
