+++ src/sparnatural/globals/statistics/initStatistics.ts
@@ -1,4 +1,4 @@
-import { getSettings } from "../../../configs/client-configs/settings";
+import { getSettings } from "../../../configs/client-configs/defaultSettings";
 import { FilteringSpecificationProvider } from "../../spec-providers/FilteringSpecificationProvider";
 import { SimpleStatisticsHandler } from "./StatisticsHandlers";
